+++ kaithem/src/cliserver.py
@@ -4,6 +4,7 @@
 import base64
 import getpass
 import hmac
+import importlib
 import logging
 import os
 import weakref
@@ -30,20 +31,20 @@ secret_key = base64.b64encode(os.urandom(24)).decode()
 with open(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", "w") as f:
     pass
 
-os.chmod(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", 0o500)
+os.chmod(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", 0o600)
 
 with open(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", "a") as f:
     f.write(secret_key)
 
 # be defensive
-os.chmod(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", 0o500)
+os.chmod(f"/dev/shm/kaithem-api-key-{getpass.getuser()}", 0o600)
 
 
 with open(f"/dev/shm/kaithem-api-port-{getpass.getuser()}", "w") as f:
     f.write(str(config.config["http_port"]))
 
 
-class WebDevices:
+class WebAPI:
     @cherrypy.expose
     def cmd(self, cmd, *args, api_key: str = "", **kw):
         if not hmac.compare_digest(api_key, secret_key):
@@ -51,3 +52,13 @@ class WebDevices:
 
         o = commands[cmd]
         o.run(*args, *kw)
+
+
+class CallAribitraryFunctionCommand(Command):
+    def run(self, module, f, *args, **kw):
+        """Usage: kaithem-cli from modulename functionname arg1 arg2 arg3"""
+        mo = importlib.import_module(module)
+        return getattr(mo, f)(*args, **kw)
+
+
+commands["from"] = CallAribitraryFunctionCommand()