+++ kaithem/src/resource_types.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: Copyright Daniel Dunn
+# SPDX-License-Identifier: GPL-3.0-only
+
 import html
 import traceback
 import weakref